replace with float

+++ CS/BlazorGridUnboundColumns/Data/EditableWeatherForecast.cs
@@ -1,8 +1,8 @@
 ﻿namespace BlazorGridUnboundColumns.Data {
     public class EditableWeatherForecast : WeatherForecast {
-        public double TemperatureF {
-            get => (base.TemperatureC / 0.5556) + 32;
-            set => base.TemperatureC = (value - 32) * 0.5556;
+        public float TemperatureF {
+            get => (float)((base.TemperatureC / 0.5556) + 32);
+            set => base.TemperatureC = (float)((value - 32) * 0.5556);
         }
     }
 }
